[rename] utils.ts to styles.ts
